refactor(move): Extract out standard path method into driver io module. Update cli to import from new module location.
