+++ fhy/cli.py
@@ -12,9 +12,9 @@ from fhy.driver.ast_program_builder.source_file_ast import (
     SourceFileAST,
     build_source_file_ast,
 )
+from fhy.driver.file_reader import standard_path
 from fhy.lang.ast.pprint import pformat_ast
 from fhy.lang.ast.serialization.to_json import dump
-from fhy.utils.discovery import standard_path
 from fhy.utils.logger import add_file_handler, get_logger
